refactor: separate backward-compatibility code into another wrapper
